first steps on relative focuser and zwo

+++ devices/sequencer.py
@@ -284,12 +284,13 @@ class Sequencer:
                     # this happens with theskyx
                     # Check if the directory is empty
                     if not os.listdir(directories[q]):
-                        # Remove the empty directory
-                        try:
-                            os.rmdir(directories[q])
-                            plog(f"The directory {directories[q]} was empty and has been removed.")
-                        except:
-                            pass
+                        if 'calibmasters' not in directories[q]:
+                            # Remove the empty directory
+                            try:
+                                os.rmdir(directories[q])
+                                plog(f"The directory {directories[q]} was empty and has been removed.")
+                            except:
+                                pass
 
                     # else:
                     #     p(f"The directory {directories[q]} is not empty.")
@@ -1619,7 +1620,7 @@ class Sequencer:
                                 except:
                                     plog(traceback.format_exc())
                                     if g_dev['mnt'].theskyx:
-                                        self.kill_and_reboot_theskyx(new_ra, new_dec)
+                                        g_dev['obs'].kill_and_reboot_theskyx(new_ra, new_dec)
                                     else:
                                         plog(traceback.format_exc())
                                 g_dev['mnt'].wait_for_slew(wait_after_slew=False)
@@ -1629,7 +1630,7 @@ class Sequencer:
                                     plog("The SkyX had an error.")
                                     plog("Usually this is because of a broken connection.")
                                     plog("Killing then waiting 60 seconds then reconnecting")
-                                    self.kill_and_reboot_theskyx(new_ra,new_dec)
+                                    g_dev['obs'].kill_and_reboot_theskyx(new_ra,new_dec)
 
                             g_dev['mnt'].wait_for_slew(wait_after_slew=False)
                             # try:
@@ -2325,7 +2326,7 @@ class Sequencer:
         # Daily reboot of necessary windows 32 programs *Cough* Theskyx *Cough*
         if g_dev['mnt'].theskyx: # It is only the mount that is the reason theskyx needs to reset
             plog ("Got here")
-            self.kill_and_reboot_theskyx(-1,-1)
+            g_dev['obs'].kill_and_reboot_theskyx(-1,-1)
             plog ("But didn't get here")
         return
 
@@ -5503,7 +5504,7 @@ class Sequencer:
                 plog ("Difficulty in directly slewing to object")
                 plog(traceback.format_exc())
                 if g_dev['mnt'].theskyx:
-                    self.kill_and_reboot_theskyx(grid_star[0]/15, grid_star[1])
+                    g_dev['obs'].kill_and_reboot_theskyx(grid_star[0]/15, grid_star[1])
                 else:
                     plog(traceback.format_exc())
 
@@ -5781,7 +5782,7 @@ class Sequencer:
                 plog ("Difficulty in directly slewing to object")
                 plog(traceback.format_exc())
                 if g_dev['mnt'].theskyx:
-                    self.kill_and_reboot_theskyx(grid_star[0] / 15, grid_star[1])
+                    g_dev['obs'].kill_and_reboot_theskyx(grid_star[0] / 15, grid_star[1])
                 else:
                     plog(traceback.format_exc())
 
@@ -6210,7 +6211,7 @@ class Sequencer:
                 except:
                     plog(traceback.format_exc())
                     if g_dev['mnt'].theskyx:
-                        self.kill_and_reboot_theskyx(g_dev["mnt"].last_ra_requested, g_dev["mnt"].last_dec_requested)
+                        g_dev['obs'].kill_and_reboot_theskyx(g_dev["mnt"].last_ra_requested, g_dev["mnt"].last_dec_requested)
                     else:
                         plog(traceback.format_exc())
